requirements.txt
flask==2.1.1
flask-sqlalchemy==2.5.1
requests==2.27.1